feat: refactor family view (#263)
