+++ 14th-team5-iOS/App/Sources/Presentation/Home/Dependency/HomeDIContainer.swift
@@ -44,7 +44,7 @@ public final class HomeDIContainer {
     }
     
     public func makeReactor() -> Reactor {
-        return HomeViewReactor(familyRepository: makeFamilyUseCase(), postRepository: makePostUseCase(), familyUseCase: makeInviteFamilyUseCase())
+        return HomeViewReactor(postRepository: makePostUseCase())
     }
     
 }
