+++ fastlane/.env.rtf
@@ -1,0 +1,20 @@
+APP_NAME=Bibbi
+WIDGET_NAME=WidgetExtension
+PRD_SCHEME=App-PRD
+DEV_SCHEME=App
+BUNDLE_ID=com.5ing.bibbi
+WIDGET_BUNDLE_ID=com.5ing.bibbi.widget
+
+APPLE_ID=55ing.team@gmail.com
+APPLE_PASSWORD=1g2ZdeMJw7FVvS
+
+APP_STORE_CONNECT_API_KEY_KEY_ID=V74YKCCUMW
+APP_STORE_CONNECT_API_KEY_ISSUER_ID=0337f6ec-46b1-488f-8868-ca7e96dea2b0
+APP_STORE_CONNECT_API_KEY_KEY=LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JR1RBZ0VBTUJNR0J5cUdTTTQ5QWdFR0NDcUdTTTQ5QXdFSEJIa3dkd0lCQVFRZ1ZMaXZBOVRLYWxQV0owK2sKbE1saDJiQURJN0ovUmFhVnhVNW5LLzhGc1AyZ0NnWUlLb1pJemowREFRZWhSQU5DQUFSRGtlR0JOcENhOEowRwpSZlVuOXpuRHJnT1FyN2g0Z2RIUjI1ZWt5RnlhYVZ6dHpwTXZRYklMMlFHZVptVnFhbndzZDQvbjl0aVRBTk9LClh4b1gyTERiCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0=
+
+KEYCHAIN_NAME=BibbiKeychain
+KEYCHAIN_PASSWORD=1g2ZdeMJw7FVvS
+
+SLACK_HOOK_URL=https://hooks.slack.com/services/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
+
+PROJECT_PATH=./14th-team5-iOS/App/App.xcodeproj
